+++ gcp/resource_netapp_gcp_volume_test.go
@@ -145,7 +145,6 @@ func testAccVolumeConfigCreate() string {
 		provider = netapp-gcp
 		name = "terraform-acceptance-test-1"
 		region = "us-east4"
-		zone = "gcp-zone"
 		storage_class = "hardware"
 		protocol_types = ["NFSv3"]
 		network = "cvs-terraform-vpc"
@@ -208,7 +207,6 @@ func testAccVolumeConfigUpdate() string {
 		provider = netapp-gcp
 		name = "terraform-acceptance-test-1"
 		region = "us-east4"
-		zone = "gcp-zone"
 		storage_class = "hardware"
 		protocol_types = ["NFSv3"]
 		network = "cvs-terraform-vpc"